Create README.md
Hi, my name is Cristian, I live in Chisinau, love to
volunteer, read, draw, watch movies, and
more. I am 19, and I enjoy meeting new people
and finding ways to interact, energetic and
eager to learn new things, help them, all while
sharing and enjoying an uplifting experience,
and try every I can to do 101% using a creative
approach to it. I love tech, IT, electronics in
general. I’m studying Software Engineering at
UTM, and I do enjoy it. This year there is a lot of
hard work and a lot of reading and memorizing,
but I think that I need to adapt faster, and
continue to absorb all the knowledge and even
more.